refactor: split main sections into components
